+++ src/Bundle/Menu/Admin/AbstractMenuNodeAdmin.php
@@ -3,12 +3,11 @@
 /*
  * This file is part of the Symfony CMF package.
  *
- * (c) 2011-2013 Symfony CMF
+ * (c) 2011-2014 Symfony CMF
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-
 
 namespace Symfony\Cmf\Bundle\MenuBundle\Admin;
 
@@ -93,27 +92,6 @@ abstract class AbstractMenuNodeAdmin extends Admin
             ->add('uri', 'text')
             ->add('content', 'text')
         ;
-    }
-
-    /**
-     * @return MenuNode
-     */
-    public function getNewInstance()
-    {
-        /** @var $new MenuNode */
-        $new = parent::getNewInstance();
-
-        if ($this->hasRequest()) {
-
-            // Set the parent
-            $parentId = $this->getRequest()->query->get('parent');
-
-            if (null !== $parentId) {
-                $new->setParent($this->getModelManager()->find(null, $parentId));
-            }
-        }
-
-        return $new;
     }
 
     public function getExportFormats()