+++ src/Bundle/Menu/Doctrine/Phpcr/MenuNodeBase.php
@@ -3,17 +3,16 @@
 /*
  * This file is part of the Symfony CMF package.
  *
- * (c) 2011-2013 Symfony CMF
+ * (c) 2011-2014 Symfony CMF
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
 
-
 namespace Symfony\Cmf\Bundle\MenuBundle\Doctrine\Phpcr;
 
-use Symfony\Cmf\Bundle\MenuBundle\Model\MenuNodeBase as BaseMenuNodeBase;
+use Symfony\Cmf\Bundle\MenuBundle\Model\MenuNodeBase as ModelMenuNodeBase;
 
-class MenuNodeBase extends BaseMenuNodeBase
+class MenuNodeBase extends ModelMenuNodeBase
 {
 }