+++ src/Bundle/Menu/Tests/Functional/Doctrine/Phpcr/MenuNodeTest.php
@@ -3,31 +3,41 @@
 /*
  * This file is part of the Symfony CMF package.
  *
- * (c) 2011-2013 Symfony CMF
+ * (c) 2011-2014 Symfony CMF
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
 
-
 namespace Symfony\Cmf\Bundle\MenuBundle\Tests\Functional\Doctrine\Phpcr;
 
+use Doctrine\ODM\PHPCR\DocumentManager;
+use Symfony\Cmf\Bundle\MenuBundle\Tests\Resources\Document\Content;
 use Symfony\Cmf\Component\Testing\Functional\BaseTestCase;
 use Symfony\Cmf\Bundle\MenuBundle\Doctrine\Phpcr\MenuNode;
-use Symfony\Cmf\Component\Testing\Document\Content;
 
 class MenuNodeTest extends BaseTestCase
 {
+    private $content;
+    /**
+     * @var DocumentManager
+     */
+    private $dm;
+    private $rootDocument;
+    /**
+     * @var MenuNode
+     */
+    private $child1;
+
     public function setUp()
     {
         $this->db('PHPCR')->createTestNode();
         $this->dm = $this->db('PHPCR')->getOm();
-        $this->baseNode = $this->dm->find(null, '/test');
+        $this->rootDocument = $this->dm->find(null, '/test');
 
-        $this->content = new Content;
-        $this->content->setParent($this->baseNode);
+        $this->content = new Content();
+        $this->content->setId('/test/fake_weak_content');
         $this->content->setTitle('fake_weak_content');
-        $this->content->setName('fake_weak_content');
         $this->dm->persist($this->content);
 
         $this->child1 = new MenuNode;
@@ -75,10 +85,13 @@ class MenuNodeTest extends BaseTestCase
             'displayChildren' => false,
         );
 
+        $startDateString = $data['publishStartDate']->format('Y-m-d');
+        $endDateString = $data['publishEndDate']->format('Y-m-d');
+
         $menuNode = new MenuNode;
         $refl = new \ReflectionClass($menuNode);
 
-        $menuNode->setParent($this->baseNode);
+        $menuNode->setParentDocument($this->rootDocument);
 
         foreach ($data as $key => $value) {
             $refl = new \ReflectionClass($menuNode);
@@ -122,8 +135,8 @@ class MenuNodeTest extends BaseTestCase
 
         $this->assertInstanceOf('\DateTime', $publishStartDate);
         $this->assertInstanceOf('\DateTime', $publishEndDate);
-        $this->assertEquals($data['publishStartDate']->format('Y-m-d'), $publishStartDate->format('Y-m-d'));
-        $this->assertEquals($data['publishEndDate']->format('Y-m-d'), $publishEndDate->format('Y-m-d'));
+        $this->assertEquals($startDateString, $publishStartDate->format('Y-m-d'));
+        $this->assertEquals($endDateString, $publishEndDate->format('Y-m-d'));
 
         // test multi-lang
         $menuNode->setLocale('fr');
@@ -133,5 +146,14 @@ class MenuNodeTest extends BaseTestCase
 
         $menuNode = $this->dm->findTranslation(null, '/test/test-node', 'fr');
         $this->assertEquals('fr', $menuNode->getLocale());
+
+        $child = $this->dm->find(null, '/test/test-node/child1');
+        $menuNode = $child->getParent();
+        $this->assertCount(1, $menuNode->getChildren());
+        $menuNode->removeChild($child);
+        $this->dm->flush();
+        $this->dm->clear();
+        $menuNode = $this->dm->find(null, '/test/test-node');
+        $this->assertCount(0, $menuNode->getChildren());
     }
 }