+++ src/Bundle/Menu/Tests/Resources/Controller/VoterController.php
@@ -1,5 +1,14 @@
 <?php
 
+/*
+ * This file is part of the Symfony CMF package.
+ *
+ * (c) 2011-2014 Symfony CMF
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace Symfony\Cmf\Bundle\MenuBundle\Tests\Resources\Controller;
 
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
@@ -49,6 +58,7 @@ class VoterController extends Controller
     public function postAction(Request $request)
     {
         $content = $request->get(DynamicRouter::CONTENT_KEY);
+
         return $this->render('::tests/voter/post.html.twig', array('content' => $content));
     }