+++ src/Bundle/Menu/Tests/Resources/DataFixtures/PHPCR/LoadMenuData.php
@@ -3,44 +3,31 @@
 /*
  * This file is part of the Symfony CMF package.
  *
- * (c) 2011-2013 Symfony CMF
+ * (c) 2011-2014 Symfony CMF
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
 
-
 namespace Symfony\Cmf\Bundle\MenuBundle\Tests\Resources\DataFixtures\PHPCR;
 
 use Doctrine\Common\DataFixtures\FixtureInterface;
 use Doctrine\Common\Persistence\ObjectManager;
-use Doctrine\Common\DataFixtures\DependentFixtureInterface;
 use Doctrine\ODM\PHPCR\DocumentManager;
 use Symfony\Cmf\Bundle\MenuBundle\Doctrine\Phpcr\MenuNode;
 use Symfony\Cmf\Bundle\MenuBundle\Doctrine\Phpcr\Menu;
 use Symfony\Cmf\Bundle\MenuBundle\Tests\Resources\Document\Content;
 use Symfony\Cmf\Bundle\MenuBundle\Tests\Resources\Document\Post;
-use Doctrine\ODM\PHPCR\Document\Generic;
 use Symfony\Cmf\Bundle\RoutingBundle\Doctrine\Phpcr\Route;
 use PHPCR\Util\NodeHelper;
 
-class LoadMenuData implements FixtureInterface, DependentFixtureInterface
+class LoadMenuData implements FixtureInterface
 {
-    protected $root;
     protected $menuRoot;
     protected $routeRoot;
 
-    public function getDependencies()
-    {
-        return array(
-            'Symfony\Cmf\Component\Testing\DataFixtures\PHPCR\LoadBaseData',
-        );
-    }
-
     public function load(ObjectManager $manager)
     {
-        $this->root = $manager->find(null, '/test');
-
         NodeHelper::createPath($manager->getPhpcrSession(), '/test/menus');
         NodeHelper::createPath($manager->getPhpcrSession(), '/test/routes/contents');
         $this->menuRoot = $manager->find(null, '/test/menus');
@@ -61,11 +48,11 @@ class LoadMenuData implements FixtureInterface, DependentFixtureInterface
         $menu = new Menu;
         $menu->setName('test-menu');
         $menu->setLabel('Test Menu');
-        $menu->setParent($this->menuRoot);
+        $menu->setParentDocument($this->menuRoot);
         $manager->persist($menu);
 
         $menuNode = new MenuNode;
-        $menuNode->setParent($menu);
+        $menuNode->setParentDocument($menu);
         $menuNode->setLabel('item-1');
         $menuNode->setName('item-1');
         $manager->persist($menuNode);
@@ -73,7 +60,7 @@ class LoadMenuData implements FixtureInterface, DependentFixtureInterface
         $content->addMenuNode($menuNode);
 
         $menuNode = new MenuNode;
-        $menuNode->setParent($menu);
+        $menuNode->setParentDocument($menu);
         $menuNode->setLabel('This node has a URI');
         $menuNode->setUri('http://www.example.com');
         $menuNode->setName('item-2');
@@ -82,20 +69,20 @@ class LoadMenuData implements FixtureInterface, DependentFixtureInterface
         $content->addMenuNode($menuNode);
 
         $subNode = new MenuNode;
-        $subNode->setParent($menuNode);
+        $subNode->setParentDocument($menuNode);
         $subNode->setLabel('@todo this node should have content');
         $subNode->setName('sub-item-1');
         $manager->persist($subNode);
 
         $subNode = new MenuNode;
-        $subNode->setParent($menuNode);
+        $subNode->setParentDocument($menuNode);
         $subNode->setLabel('This node has an assigned route');
         $subNode->setName('sub-item-2');
         $subNode->setRoute('link_test_route');
         $manager->persist($subNode);
 
         $subNode = new MenuNode;
-        $subNode->setParent($menuNode);
+        $subNode->setParentDocument($menuNode);
         $subNode->setLabel('This node has an assigned route with parameters');
         $subNode->setName('sub-item-3');
         $subNode->setRoute('link_test_route_with_params');
@@ -103,7 +90,7 @@ class LoadMenuData implements FixtureInterface, DependentFixtureInterface
         $manager->persist($subNode);
 
         $menuNode = new MenuNode;
-        $menuNode->setParent($menu);
+        $menuNode->setParentDocument($menu);
         $menuNode->setLabel('item-3');
         $menuNode->setName('item-3');
         $manager->persist($menuNode);
@@ -111,11 +98,11 @@ class LoadMenuData implements FixtureInterface, DependentFixtureInterface
         $menu = new Menu;
         $menu->setName('another-menu');
         $menu->setLabel('Another Menu');
-        $menu->setParent($this->menuRoot);
+        $menu->setParentDocument($this->menuRoot);
         $manager->persist($menu);
 
         $menuNode = new MenuNode;
-        $menuNode->setParent($menu);
+        $menuNode->setParentDocument($menu);
         $menuNode->setLabel('This node has uri, route and content set. but linkType is set to route');
         $menuNode->setLinkType('route');
         $menuNode->setUri('http://www.example.com');
@@ -124,7 +111,7 @@ class LoadMenuData implements FixtureInterface, DependentFixtureInterface
         $manager->persist($menuNode);
 
         $menuNode = new MenuNode;
-        $menuNode->setParent($menu);
+        $menuNode->setParentDocument($menu);
         $menuNode->setLabel('item-2');
         $menuNode->setName('item-2');
         $manager->persist($menuNode);
@@ -202,32 +189,32 @@ class LoadMenuData implements FixtureInterface, DependentFixtureInterface
         $menu = new Menu;
         $menu->setName('side-menu');
         $menu->setLabel('Side Menu');
-        $menu->setParent($this->menuRoot);
+        $menu->setParentDocument($this->menuRoot);
         $manager->persist($menu);
 
         $menuNode = new MenuNode;
-        $menuNode->setParent($menu);
+        $menuNode->setParentDocument($menu);
         $menuNode->setLabel('Default Behavior');
         $menuNode->setName('default');
         $menuNode->setRoute('current_menu_item_default');
         $manager->persist($menuNode);
 
         $menuNode = new MenuNode;
-        $menuNode->setParent($menu);
+        $menuNode->setParentDocument($menu);
         $menuNode->setLabel('Request Content Identity Voter');
         $menuNode->setName('request-content-identity-voter');
         $menuNode->setContent($content);
         $manager->persist($menuNode);
 
         $menuNode = new MenuNode;
-        $menuNode->setParent($menu);
+        $menuNode->setParentDocument($menu);
         $menuNode->setLabel('URI Prefix Voter');
         $menuNode->setName('uri-prefix-voter');
         $menuNode->setContent($articlesRoute);
         $manager->persist($menuNode);
 
         $menuNode = new MenuNode;
-        $menuNode->setParent($menu);
+        $menuNode->setParentDocument($menu);
         $menuNode->setLabel('Request Parent Content Identity Voter');
         $menuNode->setName('request-parent-content-identity-voter');
         $menuNode->setContent($blog);