+++ src/Bundle/Menu/Tests/Resources/Document/Post.php
@@ -1,15 +1,17 @@
 <?php
 
+/*
+ * This file is part of the Symfony CMF package.
+ *
+ * (c) 2011-2014 Symfony CMF
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace Symfony\Cmf\Bundle\MenuBundle\Tests\Resources\Document;
 
-use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\ODM\PHPCR\Mapping\Annotations as PHPCRODM;
-
-use Knp\Menu\NodeInterface;
-
-use Symfony\Component\Routing\Route;
-use Symfony\Cmf\Component\Routing\RouteReferrersReadInterface;
-use Symfony\Cmf\Bundle\MenuBundle\Model\MenuNodeReferrersInterface;
 
 /**
  * @PHPCRODM\Document(referenceable=true)