+++ src/Bundle/Menu/Tests/Unit/Provider/PhpcrMenuProviderTest.php
@@ -3,12 +3,11 @@
 /*
  * This file is part of the Symfony CMF package.
  *
- * (c) 2011-2013 Symfony CMF
+ * (c) 2011-2014 Symfony CMF
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-
 
 namespace Symfony\Cmf\Bundle\MenuBundle\Tests\Unit\Provider;
 
@@ -16,16 +15,37 @@ use Symfony\Cmf\Bundle\MenuBundle\Provider\PhpcrMenuProvider;
 
 class PhpcrMenuProviderTest extends \PHPUnit_Framework_Testcase
 {
+    private function getDmMock($path)
+    {
+        $session = $this->getMock('PHPCR\SessionInterface');
+        $session->expects($this->once())
+            ->method('getNode')
+            ->with($path)
+        ;
+        $dm = $this
+            ->getMockBuilder('Doctrine\ODM\PHPCR\DocumentManager')
+            ->disableOriginalConstructor()
+            ->getMock()
+        ;
+        $dm->expects($this->once())
+            ->method('getPhpcrSession')
+            ->will($this->returnValue($session))
+        ;
+
+        return $dm;
+    }
+
     /**
      *  @dataProvider getMenuTests
      */
     public function testGet($menuRoot, $name, $expectedPath)
     {
-        $objectManager = $this->getMock('Doctrine\Common\Persistence\ObjectManager');
+        $objectManager = $this->getDmMock($expectedPath);
         $objectManager->expects($this->once())
             ->method('find')
             ->with($this->equalTo(null), $this->equalTo($expectedPath))
-            ->will($this->returnValue($this->getMock('Knp\Menu\NodeInterface')));
+            ->will($this->returnValue($this->getMock('Knp\Menu\NodeInterface')))
+        ;
 
         $managerRegistry = $this->getMock('Doctrine\Common\Persistence\ManagerRegistry');
         $managerRegistry->expects($this->once())
@@ -53,7 +73,7 @@ class PhpcrMenuProviderTest extends \PHPUnit_Framework_Testcase
      */
     public function testHas($menuRoot, $name, $expectedPath)
     {
-        $objectManager = $this->getMock('Doctrine\Common\Persistence\ObjectManager');
+        $objectManager = $this->getDmMock($expectedPath);
         $objectManager->expects($this->once())
             ->method('find')
             ->with($this->equalTo(null), $this->equalTo($expectedPath))