+++ src/Bundle/Menu/Tests/WebTest/Voter/BaseTestCase.php
@@ -1,5 +1,14 @@
 <?php
 
+/*
+ * This file is part of the Symfony CMF package.
+ *
+ * (c) 2011-2014 Symfony CMF
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace Symfony\Cmf\Bundle\MenuBundle\Tests\WebTest\Voter;
 
 use Symfony\Component\DomCrawler\Crawler;
@@ -18,17 +27,8 @@ abstract class BaseTestCase extends BaseBaseTestCase
     protected function assertCurrentItem(Crawler $crawler, $title)
     {
         $res = $crawler->filter('li.current:contains("'.$title.'")')->count();
-        if (method_exists($crawler, 'html')) {
-            // since symfony 2.3
-            $html = $crawler->html();
-        } else {
-            // symfony 2.2
-            $html = '';
-            foreach ($crawler as $domElement) {
-                $html .= $domElement->ownerDocument->saveHTML($domElement);
-            }
+        $html = $crawler->html();
 
-        }
         $this->assertEquals(1, $res, 'Failed matching current menu item "'.$title.'", got '.$html);
     }
 }