+++ src/Bundle/Menu/Tests/WebTest/Voter/RequestContentIdentityVoterTest.php
@@ -1,8 +1,17 @@
 <?php
 
+/*
+ * This file is part of the Symfony CMF package.
+ *
+ * (c) 2011-2014 Symfony CMF
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace Symfony\Cmf\Bundle\MenuBundle\Tests\WebTest\Voter;
 
-class RequestContentIdentityTest extends BaseTestCase
+class RequestContentIdentityVoterTest extends BaseTestCase
 {
     public function testRequestContentIdentityVoterNoOp()
     {
@@ -26,4 +35,3 @@ class RequestContentIdentityTest extends BaseTestCase
         $this->assertEquals(200, $res->getStatusCode());
     }
 }
-