+++ src/Bundle/Menu/Tests/WebTest/Voter/RequestParentContentIdentityVoterTest.php
@@ -1,12 +1,21 @@
 <?php
 
+/*
+ * This file is part of the Symfony CMF package.
+ *
+ * (c) 2011-2014 Symfony CMF
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace Symfony\Cmf\Bundle\MenuBundle\Tests\WebTest\Voter;
 
-class RequestContentParentIdentityTest extends BaseTestCase
+class RequestParentContentIdentityVoterTest extends BaseTestCase
 {
     public function testRequestContentParentIdentityNoOp()
     {
-        // this test loads the "blog" page which corresponds directly 
+        // this test loads the "blog" page which corresponds directly
         // to the "Request Content PArent Identity" menu item and so DOES NOT invoke
         // the voter.
         $crawler = $this->client->request('GET', '/blog');
@@ -24,4 +33,3 @@ class RequestContentParentIdentityTest extends BaseTestCase
         $this->assertEquals(200, $res->getStatusCode());
     }
 }
-