+++ src/Bundle/Menu/Tests/WebTest/Voter/UriPrefixVoterTest.php
@@ -1,12 +1,21 @@
 <?php
 
+/*
+ * This file is part of the Symfony CMF package.
+ *
+ * (c) 2011-2014 Symfony CMF
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace Symfony\Cmf\Bundle\MenuBundle\Tests\WebTest\Voter;
 
 class UriPrefixVoterTest extends BaseTestCase
 {
     public function testUriPrefixArticlesHomepage()
     {
-        // this test loads the "articles" page which corresponds directly 
+        // this test loads the "articles" page which corresponds directly
         // to the "URI Prefix Voter" menu item and so DOES NOT invoke
         // the voter.
         $crawler = $this->client->request('GET', '/articles');
@@ -25,4 +34,3 @@ class UriPrefixVoterTest extends BaseTestCase
         $this->assertEquals(200, $res->getStatusCode());
     }
 }
-