+++ src/Bundle/Menu/Voter/RequestParentContentIdentityVoter.php
@@ -3,12 +3,11 @@
 /*
  * This file is part of the Symfony CMF package.
  *
- * (c) 2011-2013 Symfony CMF
+ * (c) 2011-2014 Symfony CMF
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-
 
 namespace Symfony\Cmf\Bundle\MenuBundle\Voter;
 
@@ -18,7 +17,7 @@ use Symfony\Component\HttpFoundation\Request;
 /**
  * A variant of the RequestContentIdentityVoter that checks if the request
  * content is of a specific class and if so checks if the value returned by
- * *getParent()* is identical to the content item in the menu items extras.
+ * *getParentDocument()* is identical to the content item in the menu items extras.
  *
  * Note that there is no check, you have to make sure the $childClass does
  * indeed have a getParent method.
@@ -74,7 +73,7 @@ class RequestParentContentIdentityVoter implements VoterInterface
         if (null !== $content
             && $this->request->attributes->has($this->requestKey)
             && $this->request->attributes->get($this->requestKey) instanceof $this->childClass
-            && $this->request->attributes->get($this->requestKey)->getParent() === $content
+            && $this->request->attributes->get($this->requestKey)->getParentDocument() === $content
         ) {
             return true;
         }